# apps/web/app/lib/audit-types.ts
// Generated by scripts/export_audit_types.py. Do not edit.

export type CreateAuditRequest = {
  repository_ids: Array<string>;
  deep_review?: boolean;
  force?: boolean;
};

export type PlanDecisionRequest = {
  version: string;
  decision: "approve" | "reject" | "revise";
  feedback?: string;
};

export type RepositoryAuditResult = {
  id: string;
  repository_id: string;
  repository: string;
  stage: string;
  base_sha: string | null;
  checks: Array<Record<string, unknown>>;
  findings: Array<Record<string, unknown>>;
  validation: Array<Record<string, unknown>>;
  plan: Record<string, unknown> | null;
  plan_version: string | null;
  approved_version: string | null;
  session_name: string | null;
  session_purpose: string;
  remote_state: string | null;
  corrections: number;
  message: string | null;
  created_at: string;
  completed_at: string | null;
  pr_url: string | null;
};

export type AuditSnapshot = {
  id: string;
  created_at: string;
  stopped: boolean;
  deep_review: boolean;
  force: boolean;
  status: string;
  summary: Record<string, number>;
  repositories: Array<RepositoryAuditResult>;
};